backend/Project_Content/requirements.txt
bcrypt==4.0.1
cryptography
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
python-multipart
passlib[bcrypt]==1.7.4
pydantic
email-validator
python-dotenv
pytest
pytest-asyncio
httpx
aiofiles
jinja2
boto3
requests
moviepy==1.0.3
srt

# OpenAI for AI content generation
openai>=1.0.0

# Firebase
firebase-admin==6.4.0
google-cloud-firestore>=2.11.0

# HTTP client for Meta API
httpx>=0.25.0

